regroupement/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

PyYAML==5.3.1
gensim==3.8.0
joblib==0.13.2

matplotlib==3.1.0

multi_rake==0.0.1
nltk==3.3
numpy==1.14.3
pandas==0.23.0
plotly==4.10.0
pyLDAvis==2.1.2
scipy==1.1.0
spacy==2.1.4
https://github.com/explosion/spacy-models/releases/download/fr_core_news_md-2.1.0/fr_core_news_md-2.1.0.tar.gz
utils==1.0.1
wordcloud==1.5.0
